th-api/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
pycryptodome==3.18.0
python-dotenv==1.0.0
requests==2.31.0
tinytuya==1.12.11
urllib3==2.0.4
